Create new version-properties.sh script
Create a new version-properties.sh script that will be used on slave
hosts to determine version information from git. This is mostly a copy
of the maven-properties.sh script. The reason we are not simply renaming
is that this commit must be merged then all slave images updated before
we change our jenkins jobs to use the new script. Renaming doesn't allow
us to do that.

Change-Id: I37c0b7cc3ca034addf32897a5a8c6501dedf00b0