Fix Jenkins reverse proxy according to docs
Jenkins has documented[0] the appropriate way to reverse proxy Jenkins
and it is complaining that our config is broken so update it according
to their docs.

Big changes here are to set nodecode on AllowEncodedSlashes and nocanon
on ProxyPass so that the PATH_INFO value is left alone. But also set
some X-Forward headers to tell Jenkins about the original port and
protocol.

Also remove the modrewrite rules so that they do not interfere with
ProxyPass as they appear to no longer be needed to host zuul status.

[0] https://wiki.jenkins-ci.org/display/JENKINS/Running+Jenkins+behind+Apache

Change-Id: I9e2b159c1bde1c55779b7b519969b167ec788ea9