add install method option for jenkins to pin
We can't sem to pin jenkins with the current
method of installing jenkins in master.pp
and per clarkb this is currently done manually.

Proposing we add an alternate method to use
dpkg to install the specific version of jenkins.

Trying to avoid changing current behavior
from defaults, so still making apt the default.

We're trying simillar implementation here:
https://review.forj.io/#/c/2549/

Change-Id: Ia2f625f52970a4c738c87d799967d373a8c2ec0e