User versioncmp to compare strings
Use versioncmp, otherwise with puppet 4 (as on Fedora 22) you get "A
String is not comparable to a non String".

We also inverted the versioncmp check for the version variable, the
main reason for this was to reduce complexity of the versioncmp code.

Additionally, defining 0 to be the latest version didn't really make
sense.  So, changing it to use 'latest' means better code readability.

Change-Id: Ic8b0fd2a19cbb90e1d57e3025a3d2513ebada2e2
Signed-off-by: Paul Belanger <pabelanger@redhat.com>