Add creation of user in jjb
This feature is useful when using unix authentication system on
jenkins. We rely on this user on the ini template, but manifest
is not creating it. So ensure that this user exists or create if
not.

As jenkins may use another types of authentication, this is
not useful in all cases. So adding a flag to allow to parameterize
it , and not creating the user by default.

Change-Id: I2a007307218a838f90c4df59ea85ad2f23c47175