Send config file to jjb command
jjb is not querying plugin info if no config file
is set. This seems to be a feature, because there are
even tests in jjb ensuring that behaviour. We call
jjb update without config file, relying on defaults,
so jenkins is not querying for plugin versions, causing
unexpected behaviours in severla plugins.

Change-Id: Ia0e110e04fbd81402ad74b94cda14f3f76f4def2