Make Jenkins vhost a host-based
Currently, when Apache2 is deciding on which vhost to service the requests,
it first looks for the VirtualHost that has the same IP and port as in the request.
(http://httpd.apache.org/docs/current/vhosts/details.html)
This means, providing "jenkins::master::vhost_name" is resolved on host,
and the host has only one IP address, that all requests match this vhost.

This prevents from hosting Jenkins on CI host together with zuul and logserver.

Change-Id: Idf16c2925ab700107a6fd311e276dc1cdb8ad44d