Allow setting java heap size in jenkins default
This is useful for downstream projects where it is needed to tune
java heap memory consumption. Pass the requiered heap size to the
jenkins class and the class set the requiered size in
jenkins.default template.
For backward compitability, if no value is passed to the
class, we simply use the default which defined in the template.

Change-Id: Ice0e4ffa2fffd041d8bcc4a0b323ffae7ba2b0c8