Customize jenkins.default by template
Turn jenkins.default to a template. This enables to customize it
easily. This is useful for downstream projects where it is needed
to tune jenkins.default parameters. Either Pass the required
parameter to jenkins class or pass a customized jenkins.default
file location in jenkins_default.
For backward compatibility, default values are set in the class
parameters list.

Change-Id: Ice0e4ffa2fffd041d8bcc4a0b323ffae7ba2b0c8