Add a flag to disable managing apache
The apache template for configuring jenkins is really basic and doesn't
allow for a lot of configuration that people might end up requiring in a
production environment. Rather than try to enhance the template simply
allow it to be disabled so a user can specify their own.

This is generally better for seperation of duties where the jenkins role
should only configure jenkins.

Change-Id: I5274af116b4a673a7d5f83a7c1142b93f2d94f70