Fix beaker on xenial
1. Since the beaker jobs are being run on xenial, we need a special nodeset
for it, otherwise beaker gives an error:

 beaker-hostgenerator was not able to use this value as input.
 Exiting with an Error.

2. Install puppet from the Ubuntu repos rather than from
puppetlabs, since puppetlabs doesn't support puppet 3 for Xenial. For
centos we can keep the install process the same.

3. Rather than updating the individual beaker tests for xenial,
this patch deletes most of the resource-level beaker assertions. It's
really quite enough to trust that a successful puppet run means the
packages were found and the config files applied. All that's really
important besides checking that puppet was successful is checking that
the services are running.

4. Don't mess with the jenkins user. Zuul needs the jenkins user's
authorized_keys file to stay intact in order to rsync logs.

5. Finally, remove the jenkins::job_builder fixture. There's not a good
excuse for this, it just doesn't seem to work right now, perhaps as a
result of recent security updates to jenkins.

Change-Id: I244ac33943e54461c0eb849d40d32a80c161e0e3
Depends-On: If31abd1d29416dbb567c0f76529913b8327d40c1