Fix SSL in beaker tests
The snakeoil package uses the host's fqdn for the cert's CN, which means
trying to use '127.0.0.1' as the JJB URL will fail with an SSL error,
and the curl tests will also fail. JJB also doesn't have --insecure or
--cafle options, so it gets a SSL_CERTIFICATE_VERIFY_FAILED when trying
to run. Use the fqdn everywhere instead of the localhost address, and
add the snakeoil certs to the system's trusted bundle.

Change-Id: Iac97910b0d04eada62dd161341ee246a9cf3ebf8